Revert "Feature/apply boundaries sidebar router (#79)"
This reverts commit 8f9e41fb2c374460a53095fb1641f899add686e9.
