+++ src/components/sidebar/backend/procedures/getInfoCardForSidebar.ts
@@ -1,6 +1,8 @@
-import { createUserOnWorkspaceContext } from '@/server/auth/userOnWorkspaceContext'
-import { getShowableSidebarCard } from '@/server/global/getShowableSidebarCard.service'
+import { workspaceEditionFilter } from '@/components/workspaces/backend/workspacesBackendUtils'
 import { protectedProcedure } from '@/server/trpc/trpc'
+import { SidebarInfoCardType } from '@/shared/globalTypes'
+import type { PrismaClient, Workspace } from '@prisma/client'
+import { Promise } from 'bluebird'
 import { z } from 'zod'
 
 const zInput = z.object({
@@ -11,13 +13,53 @@ export const getInfoCardForSidebar = protectedProcedure
   .input(zInput)
   .query(async ({ ctx, input }) => {
     const userId = ctx.session.user.id
-    const { workspaceId } = input
 
-    const context = await createUserOnWorkspaceContext(
-      ctx.prisma,
-      workspaceId,
-      userId,
-    )
+    const workspace = await ctx.prisma.workspace.findFirst({
+      where: {
+        id: input.workspaceId,
+        ...workspaceEditionFilter(userId),
+      },
+    })
 
-    return await getShowableSidebarCard(ctx.prisma, context)
+    if (!workspace) {
+      return null
+    }
+
+    if (await shouldShowInviteMembersCard(ctx.prisma, workspace)) {
+      return {
+        show: SidebarInfoCardType.Onboarding,
+      }
+    }
+
+    return null
   })
+
+const shouldShowInviteMembersCard = async (
+  prisma: PrismaClient,
+  workspace: Workspace,
+) => {
+  const numUsersPromise = prisma.usersOnWorkspaces.count({
+    where: {
+      workspaceId: workspace.id,
+    },
+  })
+
+  const numChatsPromise = prisma.chat.count({
+    where: {
+      post: {
+        workspaceId: workspace.id,
+      },
+    },
+  })
+
+  const [numUsers, numChats] = await Promise.all([
+    numUsersPromise,
+    numChatsPromise,
+  ])
+
+  if (numUsers < 2 && numChats <= 30) {
+    return true
+  }
+
+  return false
+}
