+++ src/components/sidebar/backend/procedures/getPostsForSidebar.ts
@@ -1,4 +1,3 @@
-import { createUserOnWorkspaceContext } from '@/server/auth/userOnWorkspaceContext'
 import { prisma } from '@/server/db'
 import { getSortedPostsTree } from '@/server/posts/services/getSortedPostsTree.service'
 import { protectedProcedure } from '@/server/trpc/trpc'
@@ -12,13 +11,5 @@ export const getPostsForSidebar = protectedProcedure
   .input(zInput)
   .query(async ({ ctx, input }) => {
     const userId = ctx.session.user.id
-
-    const { workspaceId } = input
-    const context = await createUserOnWorkspaceContext(
-      ctx.prisma,
-      workspaceId,
-      userId,
-    )
-
-    return await getSortedPostsTree(prisma, context)
+    return await getSortedPostsTree(prisma, userId, input.workspaceId)
   })
