+++ src/components/sidebar/backend/procedures/updatePostSortingForSidebar.ts
@@ -1,8 +1,3 @@
-import {
-  UserOnWorkspaceContext,
-  createUserOnWorkspaceContext,
-} from '@/server/auth/userOnWorkspaceContext'
-import { scopePostByWorkspace } from '@/server/posts/postUtils'
 import { protectedProcedure } from '@/server/trpc/trpc'
 import type { PrismaClientOrTrxClient } from '@/shared/globalTypes'
 import { uniq } from 'underscore'
@@ -22,34 +17,26 @@ export const updatePostSortingForSidebar = protectedProcedure
     const { workspaceId, sortedPosts } = input
     const userId = ctx.session.user.id
 
-    const context = await createUserOnWorkspaceContext(
-      ctx.prisma,
-      workspaceId,
-      userId,
-    )
-
-    await doLegacyUpdate(ctx.prisma, context, sortedPosts)
+    await doLegacyUpdate(ctx.prisma, userId, workspaceId, sortedPosts)
     // await updatePostSorting(ctx.prisma, userId, postId, previousPostId)
   })
 
 const doLegacyUpdate = async (
   prisma: PrismaClientOrTrxClient,
-  uowContext: UserOnWorkspaceContext,
+  userId: string,
+  workspaceId: string,
   sortedPosts: string[],
 ) => {
-  const { userId, workspaceId } = uowContext
   const existingPosts = await prisma.post.findMany({
     select: {
       id: true,
     },
-    where: scopePostByWorkspace(
-      {
-        id: {
-          in: sortedPosts,
-        },
+    where: {
+      id: {
+        in: sortedPosts,
       },
       workspaceId,
-    ),
+    },
   })
 
   const existingPostIds = new Set(existingPosts.map((post) => post.id))
