--- src/server/posts/postUtils.ts
@@ -1,11 +1,0 @@
-import type { Prisma } from '@prisma/client'
-
-export const scopePostByWorkspace = (
-  whereClause: Prisma.PostWhereInput,
-  workspaceId: string,
-) => {
-  return {
-    ...whereClause,
-    workspaceId,
-  }
-}
